+++ homework_04/static/css/styles.css
@@ -1,0 +1,21 @@
+    td {
+      font-size: 20px;
+      color: blue;
+    }
+    h1 {
+      color: blue;
+      font-size: 30px;
+    }
+    p {
+      font-size: 20px;
+      color: magenta;
+      }
+    title {
+      color: green;
+      font-size: 30px;
+    }
+
+    .headmark {
+      color: blue;
+      font-size: 30px;
+    }
